requirements.txt
tensorflow
pytorch
keras
sklearn
pytorch_pretrained_bert
matplotlib
keras_bert
fire
pandas
spacy
transformers